fix: update imports after DTO move
